+++ ens-20171110/alibabacloud_ens20171110/models/_describe_instances_shrink_request.py
@@ -28,6 +28,8 @@ class DescribeInstancesShrinkRequest(DaraModel):
         status: str = None,
         tags_shrink: str = None,
         v_switch_id: str = None,
+        vpd_id: str = None,
+        vpd_vswitch_id: str = None,
     ):
         self.eip_addresses_shrink = eip_addresses_shrink
         # The region ID.
@@ -85,6 +87,8 @@ class DescribeInstancesShrinkRequest(DaraModel):
         self.tags_shrink = tags_shrink
         # The ID of the vSwitch.
         self.v_switch_id = v_switch_id
+        self.vpd_id = vpd_id
+        self.vpd_vswitch_id = vpd_vswitch_id
 
     def validate(self):
         pass
@@ -157,6 +161,12 @@ class DescribeInstancesShrinkRequest(DaraModel):
         if self.v_switch_id is not None:
             result['VSwitchId'] = self.v_switch_id
 
+        if self.vpd_id is not None:
+            result['VpdId'] = self.vpd_id
+
+        if self.vpd_vswitch_id is not None:
+            result['VpdVSwitchId'] = self.vpd_vswitch_id
+
         return result
 
     def from_map(self, m: dict = None):
@@ -223,6 +233,12 @@ class DescribeInstancesShrinkRequest(DaraModel):
 
         if m.get('VSwitchId') is not None:
             self.v_switch_id = m.get('VSwitchId')
+
+        if m.get('VpdId') is not None:
+            self.vpd_id = m.get('VpdId')
+
+        if m.get('VpdVSwitchId') is not None:
+            self.vpd_vswitch_id = m.get('VpdVSwitchId')
 
         return self